Audio/SB16: Implemented support for async I/O (disabled by default); a lot of code cleanup / reorganization to support more streams in the future [build fix]. bugref:9890
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@88773 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Devices/Audio/DevSB16.cpp
@@ -2601,9 +2601,9 @@ static void sb16StreamUpdate(PSB16STREAM pStream, PAUDMIXSINK pSink)
 
             RTCircBufAcquireReadBlock(pStream->State.pCircBuf, cbToReadFromStream, &pvSrcBuf, &cbSrcBuf);
 
-            int rc2 = AudioMixerSinkWrite(pSink, AUDMIXOP_COPY, pvSrcBuf, cbSrcBuf, &cbWritten);
+            int rc2 = AudioMixerSinkWrite(pSink, AUDMIXOP_COPY, pvSrcBuf, (uint32_t)cbSrcBuf, &cbWritten);
             AssertRC(rc2);
-            Assert(cbWritten <= cbSrcBuf);
+            Assert(cbWritten <= (uint32_t)cbSrcBuf);
 
             RTCircBufReleaseReadBlock(pStream->State.pCircBuf, cbWritten);
